+++ src/BrokerLib/dcc/NetworkDeviceTasks.h
@@ -21,6 +21,8 @@
 
 #include <dcclite_shared/Packet.h>
 
+#include "../sys/Timeouts.h"
+
 namespace dcclite::broker
 { 
 	class NetworkTask;
@@ -154,7 +156,7 @@ namespace dcclite::broker
 		private:
 			bool m_fFinished = false;
 			bool m_fFailed = false;			
-	};	
+	};
 
 	class IServoProgrammerTask
 	{
@@ -181,6 +183,16 @@ namespace dcclite::broker
 	};
 
 	typedef std::vector<uint8_t> DownloadEEPromTaskResult_t;
+
+	class INetworkDevice_TaskProvider
+	{
+		public:
+			virtual [[nodiscard]] std::shared_ptr<NetworkTask> StartDownloadEEPromTask(NetworkTask::IObserver *observer, DownloadEEPromTaskResult_t &resultsStorage) = 0;
+			virtual [[nodiscard]] std::shared_ptr<NetworkTask> StartServoTurnoutProgrammerTask(NetworkTask::IObserver *observer, Decoder &decoder) = 0;
+			virtual [[nodiscard]] std::shared_ptr<NetworkTask> StartDeviceRenameTask(NetworkTask::IObserver *observer, RName newName) = 0;
+			virtual [[nodiscard]] std::shared_ptr<NetworkTask> StartDeviceClearEEPromTask(NetworkTask::IObserver *observer) = 0;
+			virtual [[nodiscard]] std::shared_ptr<NetworkTask> StartDeviceNetworkTestTask(NetworkTask::IObserver *observer, std::chrono::milliseconds timeout = TASK_NETWORK_TEST_DEFAULT_TIMEOUT) = 0;
+	};
 
 	namespace detail
 	{